refactiring registration
