+++ backend/authorize/forms.py
@@ -1,9 +1,7 @@
 from django import forms
 from django.contrib.auth.forms import UserCreationForm
-# from django.contrib.auth.models import User
 from db.models import User
 
-# Create your forms here.
 
 class NewUserForm(UserCreationForm):
     email = forms.EmailField(required=True)
