+++ backend/authorize/urls.py
@@ -1,5 +1,8 @@
 from django.urls import path, include
 
+from . import views
+
 urlpatterns = [
-    path('', include('django.contrib.auth.urls'))
+    path('auth/', include('django.contrib.auth.urls')),
+    path('registration/', views.register_request, name='registration')
 ]
