+++ backend/authorize/views.py
@@ -1,0 +1,17 @@
+from django.shortcuts import render, redirect
+from .forms import NewUserForm
+from django.contrib.auth import login
+from django.contrib import messages
+
+
+def register_request(request):
+    if request.method == 'POST':
+        form = NewUserForm(request.POST)
+        if form.is_valid():
+            user = form.save()
+            login(request, user)
+            messages.success(request, 'Registration successful.')
+            return redirect('index')
+        messages.error(request, 'Unsuccessful registration. Invalid information.')
+    form = NewUserForm()
+    return render(request=request, template_name='registration/registration.html', context={'register_form': form})
