+++ backend/urls.py
@@ -3,6 +3,5 @@ from django.urls import include, path
 urlpatterns = [
     path('', include('backend.core.urls')),
     path('athlete/', include('backend.athlete.urls')),
-    path('auth/', include('backend.authorize.urls')),
-    path('registration/', include('backend.registration.urls')),
+    path('', include('backend.authorize.urls')),
 ]
