+++ requirements.txt
@@ -2,6 +2,7 @@ asgiref==3.5.2
 black==22.6.0
 click==8.1.3
 Django==4.0.6
+django-crispy-forms==1.14.0
 flake8==5.0.4
 isort==5.10.1
 mccabe==0.7.0
